requirements.txt
Flask==3.1.3
Werkzeug==2.3.9
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.2.0
pg8000==1.32.0
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
SQLAlchemy==2.0.21
python-dotenv==1.0.0
